backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

python-docx==1.1.0
openpyxl==3.1.2

pydantic==2.6.4
pydantic-settings==2.2.1

SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

python-multipart==0.0.9

bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

loguru==0.7.2

celery==5.3.6
redis==5.0.3
pytest==8.1.1
pytest-cov==4.1.0
httpx==0.27.0